Rename toolName.ts to toolName.web.ts
