+++ tests/e2e/client.ts
@@ -2,7 +2,7 @@ import { Client } from '@modelcontextprotocol/sdk/client/index.js';
 import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio.js';
 import { z } from 'zod';
 
-import { ToolName } from '../../src/tools/toolName.js';
+import { ToolName } from '../../src/tools/toolName.web.js';
 import invariant from '../../src/utils/invariant.js';
 import { getDefaultEnv } from '../testEnv.js';
